requirements.txt
Flask
gunicorn
werkzeug
nltk
flask-cors
python-docx
Pillow
